Trait icu_datetime::date::DateInput

pub trait DateInput {
    fn year(&self) -> Option<Year>;

    fn month(&self) -> Option<Month>;

    fn day_of_month(&self) -> Option<DayOfMonth>;

    fn iso_weekday(&self) -> Option<IsoWeekday>;

    fn day_of_year_info(&self) -> Option<DayOfYearInfo>;
}

Representation of a formattable calendar date. Supports dates in any calendar system that uses solar days indexed by an era, year, month, and day.

All fields are optional. If a field is not present but is required when formatting, an error result will be returned from the formatter.

All data represented in DateInput should be locale-agnostic.

Required methods

fn year(&self) -> Option<Year>

Gets the era and year input.

fn month(&self) -> Option<Month>

Gets the month input.

fn day_of_month(&self) -> Option<DayOfMonth>

Gets the day input.

fn iso_weekday(&self) -> Option<IsoWeekday>

Gets the weekday input.

fn day_of_year_info(&self) -> Option<DayOfYearInfo>

Gets information on the position of the day within the year.

Loading content...