icu_datetime
Enum DateTimeError

#[non_exhaustive]
pub enum DateTimeError {

    Pattern(PatternError),
    Format(Error),
    Data(DataError),
    MissingInputField(Option<&'static str>),
    UnsupportedField(FieldSymbol),
    UnsupportedFormattingField(Field),
    UnsupportedOptions,
    PluralRules(PluralsError),
    DateTimeInput(CalendarError),
    MissingWeekdaySymbol(usize),
    MissingMonthSymbol(MonthCode),
    FixedDecimal,
    FixedDecimalFormatter(DecimalError),
    MismatchedAnyCalendar(AnyCalendarKind, Option<AnyCalendarKind>),
    MissingDateSymbols,
    MissingTimeSymbols,
    MissingOrdinalRules,
    MissingNames(Field),
    DuplicateField(Field),

}

A list of error outcomes for various operations in this module.

Re-exported as Error.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Pattern(PatternError)

An error originating from parsing a pattern.

Format(Error)

An error originating from the Write trait.

Data(DataError)

An error originating inside of the data provider.

MissingInputField(Option<&'static str>)

An error originating from a missing field in datetime input. TODO: How can we return which field was missing?

UnsupportedField(FieldSymbol)

An error originating from an unsupported field in a datetime format.

UnsupportedFormattingField(Field)

An unsupported field with a field length.

UnsupportedOptions

An error due to there being no patterns for the given options.

PluralRules(PluralsError)

An error originating from PluralRules.

DateTimeInput(CalendarError)

An error originating from DateTimeInput.

MissingWeekdaySymbol(usize)

An error originating from a missing weekday symbol in the data.

MissingMonthSymbol(MonthCode)

An error originating from a missing month symbol in the data.

FixedDecimal

The FixedDecimalFormatter is not loaded

FixedDecimalFormatter(DecimalError)

An error originating from FixedDecimalFormatter

MismatchedAnyCalendar(AnyCalendarKind, Option<AnyCalendarKind>)

An error from mixing calendar types in DateTimeFormatter

MissingDateSymbols

Missing date symbols

MissingTimeSymbols

Missing time symbols

MissingOrdinalRules

ordinal_rules must be set for PatternPlurals::MultipleVariants

MissingNames(Field)

The names for the given field are not loaded

DuplicateField(Field)

The same field occurs multiple times in a pattern or was loaded multiple times