icu_datetime::fields
Enum FieldSymbol

pub enum FieldSymbol {
    Era,
    Year(Year),
    Month(Month),
    Week(Week),
    Day(Day),
    Weekday(Weekday),
    DayPeriod(DayPeriod),
    Hour(Hour),
    Minute,
    Second(Second),
    TimeZone(TimeZone),
}

A field symbol for a date formatting pattern. Field symbols are a more granular distinction for a pattern field within the category of a field type. Examples of field types are: Year, Month, Hour. Within the Hour field type, examples of field symbols are: Hour::H12, Hour::H24. Each field symbol is represented within the date formatting pattern string by a distinct character from the set of A..Z and a..z.

Variants

Era

Era name.

Year(Year)

Year number or year name.

Month(Month)

Month number or month name.

Week(Week)

Week number or week name.

Day(Day)

Day number relative to a time period longer than a week (ex: month, year).

Weekday(Weekday)

Day number or day name relative to a week.

DayPeriod(DayPeriod)

Name of a period within a day.

Hour(Hour)

Hour number within a day, possibly with day period.

Minute

Minute number within an hour.

Second(Second)

Seconds number within a minute, including fractional seconds, or milliseconds within a day.

TimeZone(TimeZone)

Time zone as a name, a zone ID, or a ISO 8601 numerical offset.