Enum icu_datetime::fields::TimeZone

#[repr(u8)]
pub enum TimeZone {
    LowerZ,
    UpperZ,
    UpperO,
    LowerV,
    UpperV,
    LowerX,
    UpperX,
}

An enum for the possible symbols of a time zone field in a date pattern.

Variants

LowerZ

Field symbol for the specific non-location format of a time zone.

For example: “Pacific Standard Time”

This field symbol is represented by the character z in a date formatting pattern string.

For more details, see documentation on date field symbols.

UpperZ

Field symbol for any of: the ISO8601 basic format with hours, minutes and optional seconds fields, the long localized GMT format, or the ISO8601 extended format with hours, minutes and optional seconds fields.

This field symbol is represented by the character Z in a date formatting pattern string.

For more details, see documentation on date field symbols.

UpperO

Field symbol for the localized GMT format of a time zone.

For example: “GMT-07:00”

This field symbol is represented by the character O in a date formatting pattern string.

For more details, see documentation on date field symbols.

LowerV

Field symbol for the generic non-location format of a time zone.

For example: “Pacific Time”

This field symbol is represented by the character v in a date formatting pattern string.

For more details, see documentation on date field symbols.

UpperV

Field symbol for any of: the time zone id, time zone exemplar city, or generic location format.

This field symbol is represented by the character V in a date formatting pattern string.

For more details, see documentation on date field symbols.

LowerX

Field symbol for either the ISO8601 basic format or ISO8601 extended format, with an optional ISO8601 UTC indicator Z.

This field symbol is represented by the character x in a date formatting pattern string.

For more details, see documentation on date field symbols.

UpperX

Field symbol for either the ISO8601 basic format or ISO8601 extended format. This does not allow an optional ISO8601 UTC indicator Z, whereas TimeZone::LowerX allows the optional Z.

This field symbol is represented by the character X in a date formatting pattern string.

For more details, see documentation on date field symbols.