Module icu_datetime::fields::length

Structs

FieldLengthULE

Enums

FieldLength

An enum representing the length of a field within a date or time formatting pattern string, in which the pattern field is represented as a letter occurring 1 or more times in a row, ex: MMM, dd, y. See the LDML documentation in UTS 35 for more details.

FieldNumericOverrides

Various numeric overrides for datetime patterns as found in CLDR

LengthError

An error relating to the length of a field within a date pattern.

Constants

FIRST_FIXED 🔒

First index used for fixed size formats in compact FieldLength representation

FIRST_NUMERIC_OVERRIDE 🔒

First index used for numeric overrides in compact FieldLength representation