icu_datetime::fieldsets::enums

Enum CalendarPeriodFieldSet

#[non_exhaustive]
pub enum CalendarPeriodFieldSet {
    M(M),
    YM(YM),
    Y(Y),
}

An enumeration over all possible calendar period field sets.

This is a dynamic field set. For more information, see enums.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

M(M)

A standalone month, as in “January”.

YM(YM)

A month and year, as in “January 2000”.

Y(Y)

A year, as in “2000”.

Implementations

impl CalendarPeriodFieldSet

pub const ALL_DATA_MARKER_ATTRIBUTES: &'static [&'static DataMarkerAttributes]

All attributes associated with this enum.

Encoding Details

The string is based roughly on the UTS 35 symbol table with the following exceptions:

Lowercase letters are chosen where there is no ambiguity: E becomes e

Capitals are replaced with their lowercase and a number 0: M becomes m0

A single symbol is included for each component: length doesn’t matter

Time fields are encoded with their hour field only: j, h, or h0

Examples

use icu::datetime::fieldsets::enums::CalendarPeriodFieldSet as FS;
use icu_provider::DataMarkerAttributes;

assert!(FS::ALL_DATA_MARKER_ATTRIBUTES.contains(
    &DataMarkerAttributes::from_str_or_panic("m0")
));