icu_datetime::fieldsets::enums

Enum CompositeFieldSet

#[non_exhaustive]
pub enum CompositeFieldSet {
    Date(DateFieldSet),
    CalendarPeriod(CalendarPeriodFieldSet),
    Time(TimeFieldSet),
    Zone(ZoneFieldSet),
    DateTime(DateAndTimeFieldSet),
    DateZone(ZonedDateFieldSet),
    TimeZone(ZonedTimeFieldSet),
    DateTimeZone(ZonedDateAndTimeFieldSet),
}

An enum supporting all possible field sets and options.

This is a dynamic field set. For more information, see enums.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Date(DateFieldSet)

Field set for a date.

CalendarPeriod(CalendarPeriodFieldSet)

Field set for a calendar period.

Time(TimeFieldSet)

Field set for a time.

Zone(ZoneFieldSet)

Field set for a time zone.

DateTime(DateAndTimeFieldSet)

Field set for a date and a time together.

DateZone(ZonedDateFieldSet)

Field set for a date and a time zone together.

TimeZone(ZonedTimeFieldSet)

Field set for a time and a time zone together.

DateTimeZone(ZonedDateAndTimeFieldSet)

Field set for a date, a time, and a time zone together.