icu_datetime::fieldsets::enums

Enum TimeFieldSet

#[non_exhaustive]
pub enum TimeFieldSet {
    T(T),
}

An enumeration over all possible time field sets.

This is a dynamic field set. For more information, see enums.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

T(T)

A time of day.

Implementations

impl TimeFieldSet

pub const ALL_DATA_MARKER_ATTRIBUTES: &'static [&'static DataMarkerAttributes]

All attributes associated with this enum.

Encoding Details

The string is based roughly on the UTS 35 symbol table with the following exceptions:

Lowercase letters are chosen where there is no ambiguity: E becomes e

Capitals are replaced with their lowercase and a number 0: M becomes m0

A single symbol is included for each component: length doesnโt matter

Time fields are encoded with their hour field only: j, h, or h0

Examples

use icu::datetime::fieldsets::enums::TimeFieldSet as FS;
use icu_provider::DataMarkerAttributes;

assert!(FS::ALL_DATA_MARKER_ATTRIBUTES.contains(
    &DataMarkerAttributes::from_str_or_panic("j")
));

impl TimeFieldSet

pub fn with_zone<Z: ZoneMarkers>(self, zone: Z) -> Combo<Self, Z>

Associates this field set with a time zone field set.