icu_datetime

Module input

Types that can be fed to DateTimeFormatter/FixedCalendarDateTimeFormatter.

This module contains re-exports from the icu_calendar and icu_time crates.

Structs

Date

A date for a given calendar.

DateTime

A date and time for a given calendar.

Time

A representation of a time in hours, minutes, seconds, and nanoseconds

TimeZone

A CLDR time zone identity.

TimeZoneInfo

A utility type that can hold time zone information.

UtcOffset

An offset from Coordinated Universal Time (UTC).

ZonedDateTime

A date and time for a given calendar, local to a specified time zone.