Function icu_datetime::mock::parse_gregorian_from_str

pub fn parse_gregorian_from_str(
    input: &str
) -> Result<DateTime<Gregorian>, DateTimeError>

Temporary function for parsing a DateTime<Gregorian>

This utility is for easily creating dates, not a complete robust solution. The string must take a specific form of the ISO-8601 format: YYYY-MM-DDThh:mm:ss.

use icu::datetime::mock::parse_gregorian_from_str;
use icu_calendar::{DateTime, Gregorian};

let date: DateTime<Gregorian> = parse_gregorian_from_str("2020-10-14T13:21:00")
    .expect("Failed to parse a datetime.");

Optionally, fractional seconds can be specified: YYYY-MM-DDThh:mm:ss.SSS.

use icu::datetime::mock::parse_gregorian_from_str;
use icu_calendar::{DateTime, Gregorian};

let date: DateTime<Gregorian> = parse_gregorian_from_str("2020-10-14T13:21:00.101")
    .expect("Failed to parse a datetime.");
assert_eq!(u32::from(date.time.nanosecond), 101_000_000);