Struct icu_datetime::mock::time_zone::MockTimeZone

pub struct MockTimeZone {
    pub gmt_offset: GmtOffset,
    pub time_zone_id: Option<String>,
    pub metazone_id: Option<String>,
    pub time_variant: Option<TinyStr8>,
}

A temporary struct that implements TimeZoneInput and is used in tests, benchmarks and examples of this component.

Notice: Rust at the moment does not have a canonical way to represent time zones. We are introducing MockTimeZone as an example of the data necessary for ICU TimeZoneFormat to work, and we hope to work with the community to develop core date and time APIs that will work as an input for this component.

Examples

use icu::datetime::mock::time_zone::MockTimeZone;
use icu::datetime::date::GmtOffset;

let tz1 = MockTimeZone::new(
    GmtOffset::default(),
    /* time_zone_id  */ None,
    /* metazone_id   */ None,
    /* time_variaint */ None,
);

let tz2: MockTimeZone = "+05:00".parse()
    .expect("Failed to parse a time zone.");

Fields
gmt_offset: GmtOffset

The GMT offset in seconds.

time_zone_id: Option<String>

The IANA time-zone identifier

metazone_id: Option<String>

The CLDR metazone identifier

time_variant: Option<TinyStr8>

The time variant e.g. "daylight" or "standard"

Implementations

impl MockTimeZone

pub const fn new(
gmt_offset: GmtOffset,
time_zone_id: Option<String>,
metazone_id: Option<String>,
time_variant: Option<TinyStr8>
) -> Self

Creates a new MockTimeZone. A GMT offset is required, as it is used as a final fallback for formatting. The other arguments optionally allow access to more robust formats.