Struct icu_datetime::mock::zoned_datetime::MockZonedDateTime

pub struct MockZonedDateTime {
    pub datetime: MockDateTime,
    pub time_zone: MockTimeZone,
}

A temporary struct that implements ZonedDateTimeInput and is used in tests, benchmarks and examples of this component.

The composition of MockDateTime and MockTimeZone.

Notice: Rust at the moment does not have a canonical way to represent date and time. We are introducing MockZonedDateTime as an example of the data necessary for ICU ZonedDateTimeFormat to work, and we hope to work with the community to develop core date and time APIs that will work as an input for this component.

Examples

use icu::datetime::mock::datetime::MockDateTime;
use icu::datetime::mock::time_zone::MockTimeZone;
use icu::datetime::mock::zoned_datetime::MockZonedDateTime;

let dt: MockDateTime = "2020-10-14T13:21:00".parse()
    .expect("Failed to parse a datetime.");

let tz: MockTimeZone = "+05:00".parse()
    .expect("Failed to parse a time zone.");

let zdt1 = MockZonedDateTime::new(dt, tz);
let zdt2: MockZonedDateTime = "2020-10-14T13:21:00+05:00".parse()
    .expect("Failed to parse a zoned datetime.");

Fields
datetime: MockDateTime

The datetime component.

time_zone: MockTimeZone

The time zone component.

Implementations

impl MockZonedDateTime

pub const fn new(datetime: MockDateTime, time_zone: MockTimeZone) -> Self

Creates a new MockZonedDateTime from an already validated MockDateTime and MockTimeZone.