Module icu_datetime::provider::date_time

Traits for managing data needed by TypedDateTimeFormatter.

Structs

PatternSelector

Enums

GetSymbolForDayPeriodError 🔒

GetSymbolForEraError 🔒

GetSymbolForMonthError 🔒

GetSymbolForWeekdayError 🔒

MonthPlaceholderValue 🔒

Internal enum to represent the kinds of month symbols for interpolation

PatternForLengthError 🔒

Traits

DateSymbols 🔒

TimeSymbols 🔒

Functions

generic_pattern_for_date_length 🔒

Determine the appropriate Pattern for a given options::length::Date bag.

pattern_for_date_length 🔒

Determine the appropriate Pattern for a given options::length::Date bag.

pattern_for_date_length_inner 🔒

pattern_for_time_length 🔒

Determine the appropriate Pattern for a given options::length::Time bag. If a preference for an hour cycle is set, it will look look up a pattern in the time_h11_12 or time_h23_h24 provider data, and then manually modify the symbol in the pattern if needed.

pattern_for_time_length_inner 🔒