icu_datetime::provider::neo

Module marker_attrs

Helpers involving the data marker attributes used for date names.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Enums

Context

Field contexts supported in data marker attributes.

GlueType

Date, time, and time zone combinations supported in data marker attributes.

Length

Field lengths supported in data marker attribute.

PatternLength

Pattern lengths supported in data marker attributes.

Constants

ABBR

ABBR_STANDALONE

ABBR_STANDALONE_STR

ABBR_STR

NARROW

NARROW_STANDALONE

NARROW_STANDALONE_STR

NARROW_STR

NUMERIC

NUMERIC_STR

PATTERN_LONG

PATTERN_LONG12

PATTERN_LONG24

PATTERN_LONG12_STR

PATTERN_LONG24_STR

PATTERN_LONG_DT

PATTERN_LONG_DTZ

PATTERN_LONG_DTZ_STR

PATTERN_LONG_DT_STR

PATTERN_LONG_DZ

PATTERN_LONG_DZ_STR

PATTERN_LONG_STR

PATTERN_LONG_TZ

PATTERN_LONG_TZ_STR

PATTERN_MEDIUM

PATTERN_MEDIUM12

PATTERN_MEDIUM24

PATTERN_MEDIUM12_STR

PATTERN_MEDIUM24_STR

PATTERN_MEDIUM_DT

PATTERN_MEDIUM_DTZ

PATTERN_MEDIUM_DTZ_STR

PATTERN_MEDIUM_DT_STR

PATTERN_MEDIUM_DZ

PATTERN_MEDIUM_DZ_STR

PATTERN_MEDIUM_STR

PATTERN_MEDIUM_TZ

PATTERN_MEDIUM_TZ_STR

PATTERN_SHORT

PATTERN_SHORT12

PATTERN_SHORT24

PATTERN_SHORT12_STR

PATTERN_SHORT24_STR

PATTERN_SHORT_DT

PATTERN_SHORT_DTZ

PATTERN_SHORT_DTZ_STR

PATTERN_SHORT_DT_STR

PATTERN_SHORT_DZ

PATTERN_SHORT_DZ_STR

PATTERN_SHORT_STR

PATTERN_SHORT_TZ

PATTERN_SHORT_TZ_STR

SHORT

SHORT_STANDALONE

SHORT_STANDALONE_STR

SHORT_STR

WIDE

WIDE_STANDALONE

WIDE_STANDALONE_STR

WIDE_STR

Functions

name_attr_for

Creates a name data marker attribute from the enum values.

name_marker_attr_info

Parses a name data marker attribute to enum values.

pattern_marker_attr_for_glue

pattern_marker_attr_info_for_glue

Parses a pattern data marker attribute to enum values.