icu_datetime::provider::pattern

Module reference

Reference Pattern implementation intended to follow Unicode specification, and become publicly available for tooling to use for parsing/inspecting/modifying and serialization.

The runtime Pattern uses parsing/serialization from this module.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Structs

GenericPattern

A fully-owned, non-zero-copy type corresponding to GenericPattern.

Pattern

A fully-owned, non-zero-copy type corresponding to Pattern.