icu_datetime::provider::pattern::reference

Struct Pattern

pub struct Pattern { /* private fields */ }

A fully-owned, non-zero-copy type corresponding to Pattern.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Implementations

impl Pattern

pub fn into_items(self) -> Vec<PatternItem>

Convert a Pattern to a vector of pattern items.

The Pattern can be restored via the From impl.