icu_datetime::provider::pattern::runtime

Struct Pattern

pub struct Pattern<'data> {
    pub items: ZeroVec<'data, PatternItem>,
    pub metadata: PatternMetadata,
}

A raw, low-level pattern for datetime formatting.

It consists of an owned-or-borrowed list of PatternItems corresponding to either fields or literal characters.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Fields
items: ZeroVec<'data, PatternItem>

The list of PatternItems.

metadata: PatternMetadata

Pre-computed metadata about the pattern.

This field should contain the smallest time unit from the items vec. If it doesnโt, unexpected results for day periods may be encountered.

Implementations

impl Pattern<'_>

pub fn to_string(&self) -> String

Converts the given value to a String.

Under the hood, this uses an efficient Writeable implementation. However, in order to avoid allocating a string, it is more efficient to use Writeable directly.

impl Pattern<'_>

pub fn as_ref(&self) -> Pattern<'_>

Borrows a Pattern from another Pattern.