Struct icu_datetime::provider::time_zones::TimeZoneBcp47Id

pub struct TimeZoneBcp47Id(pub TinyAsciiStr<8>);

TimeZone ID in BCP47 format

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Tuple Fields
0: TinyAsciiStr<8>