icu_datetime

Module scaffold

Scaffolding traits and types for the datetime crate.

Items in this module are mostly for trait bounds. Most users should not need to reference these items in userland code.

Structs

DataPayloadWithVariables

An implementation of MaybePayload that wraps an optional DataPayload, parameterized by Variables.

DataPayloadWithVariablesBorrowed

Borrowed version of DataPayloadWithVariables.

Enums

FullDataCalMarkers

Implementation of CalMarkers that includes data for all calendars.

MaybePayloadError

An error returned by MaybePayload.

NoDataCalMarkers

Implementation of CalMarkers that includes data for no calendars.

Traits

AllAnyCalendarExternalDataMarkers

Trait to consolidate data provider markers external to this crate for datetime formatting with any calendar.

AllAnyCalendarFormattingDataMarkers

Trait to consolidate data provider markers defined by this crate for datetime formatting with any calendar.

AllAnyCalendarPatternDataMarkers

Trait to consolidate data provider markers defined by this crate for datetime skeleton patterns with any calendar.

AllFixedCalendarExternalDataMarkers

Trait to consolidate data provider markers external to this crate for datetime formatting with a fixed calendar.

AllFixedCalendarFormattingDataMarkers

Trait to consolidate data provider markers defined by this crate for datetime formatting with a fixed calendar.

AllFixedCalendarPatternDataMarkers

Trait to consolidate data provider markers defined by this crate for datetime skeleton patterns with a fixed calendar.

AllInputMarkers

Trait implemented on formattable datetime types.

CalMarkers

A collection of marker types associated with all formattable calendars.

CldrCalendar

A calendar that can be found in CLDR.

ConvertCalendar

A type that can be converted into a specific calendar system.

DateDataMarkers

A trait associating types for date formatting in any calendar (data markers only).

DateInputMarkers

A trait associating types for date formatting in any calendar (input types only).

DateTimeMarkers

A trait associating constants and types implementing various other traits required for datetime formatting.

DateTimeNamesFrom

A trait for a DateTimeNamesMarker that can be created from a more specific one, M.

DateTimeNamesMarker

Trait for a type that owns datetime names data, usually in the form of data payloads.

GetField

A type that can return a certain field T.

InFixedCalendar

An input associated with a fixed, static calendar.

InSameCalendar

An input that may be associated with a specific runtime calendar.

IntoFormattableAnyCalendar

A calendar type that is supported by DateTimeFormatter.

MaybePayload

A type that may or may not be a DataPayload and may or may not contain a value depending on the type parameter Variables.

NamesContainer

Trait that associates a container for a payload parameterized by the given variables.

TimeMarkers

A trait associating types for time formatting (input types and data markers).

TypedDateDataMarkers

A trait associating types for date formatting in a specific calendar (data markers only).

UnstableSealed

Trait marking other traits that are considered unstable and should not generally be implemented outside of the datetime crate.

ZoneMarkers

A trait associating types for time zone formatting (input types and data markers).