icu_datetime::scaffold

Trait AllAnyCalendarPatternDataMarkers

pub trait AllAnyCalendarPatternDataMarkers<FSet: DateTimeMarkers>:
    DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Buddhist>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Chinese>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Coptic>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Dangi>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Ethiopian>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Gregorian>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Hebrew>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Indian>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Hijri>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Japanese>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Persian>
    + DataProvider<<<FSet::D as DateDataMarkers>::Skel as CalMarkers<ErasedMarker<PackedPatterns<'static>>>>::Roc>
    + DataProvider<<FSet::T as TimeMarkers>::TimeSkeletonPatternsV1>
    + DataProvider<FSet::GluePatternV1>
where
    FSet::D: DateDataMarkers,
    FSet::T: TimeMarkers,
    FSet::Z: ZoneMarkers,
{ }

Trait to consolidate data provider markers defined by this crate for datetime skeleton patterns with any calendar.

This trait is implemented on all providers that support datetime skeleton patterns, including crate::provider::Baked.