icu_datetime::scaffold

Trait DateDataMarkers

pub trait DateDataMarkers: UnstableSealed {
    type Skel: CalMarkers<ErasedMarker<PackedPatterns<'static>>>;
    type Year: CalMarkers<YearNamesV1>;
    type Month: CalMarkers<MonthNamesV1>;
    type WeekdayNamesV1: DataMarker<DataStruct = LinearNames<'static>>;
}

A trait associating types for date formatting in any calendar (data markers only).

This is a sealed trait implemented on field set markers.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Associated Types

type Skel: CalMarkers<ErasedMarker<PackedPatterns<'static>>>

Cross-calendar data markers for date skeleta.

type Year: CalMarkers<YearNamesV1>

Cross-calendar data markers for year names.

type Month: CalMarkers<MonthNamesV1>

Cross-calendar data markers for month names.

type WeekdayNamesV1: DataMarker<DataStruct = LinearNames<'static>>

Marker for loading weekday names.