icu_datetime::scaffold

Trait DateTimeNamesFrom

pub trait DateTimeNamesFrom<M: DateTimeNamesMarker>: DateTimeNamesMarker {

    // Required methods
    fn map_year_names(
        other: <M::YearNames as NamesContainer<YearNamesV1, YearNameLength>>::Container,
    ) -> <Self::YearNames as NamesContainer<YearNamesV1, YearNameLength>>::Container;
    fn map_month_names(
        other: <M::MonthNames as NamesContainer<MonthNamesV1, MonthNameLength>>::Container,
    ) -> <Self::MonthNames as NamesContainer<MonthNamesV1, MonthNameLength>>::Container;
    fn map_weekday_names(
        other: <M::WeekdayNames as NamesContainer<WeekdayNamesV1, WeekdayNameLength>>::Container,
    ) -> <Self::WeekdayNames as NamesContainer<WeekdayNamesV1, WeekdayNameLength>>::Container;
    fn map_day_period_names(
        other: <M::DayPeriodNames as NamesContainer<DayPeriodNamesV1, DayPeriodNameLength>>::Container,
    ) -> <Self::DayPeriodNames as NamesContainer<DayPeriodNamesV1, DayPeriodNameLength>>::Container;
    fn map_zone_essentials(
        other: <M::ZoneEssentials as NamesContainer<TimezoneNamesEssentialsV1, ()>>::Container,
    ) -> <Self::ZoneEssentials as NamesContainer<TimezoneNamesEssentialsV1, ()>>::Container;
    fn map_zone_locations(
        other: <M::ZoneLocations as NamesContainer<TimezoneNamesLocationsOverrideV1, ()>>::Container,
    ) -> <Self::ZoneLocations as NamesContainer<TimezoneNamesLocationsOverrideV1, ()>>::Container;
    fn map_zone_locations_root(
        other: <M::ZoneLocationsRoot as NamesContainer<TimezoneNamesLocationsRootV1, ()>>::Container,
    ) -> <Self::ZoneLocationsRoot as NamesContainer<TimezoneNamesLocationsRootV1, ()>>::Container;
    fn map_zone_exemplars(
        other: <M::ZoneExemplars as NamesContainer<TimezoneNamesCitiesOverrideV1, ()>>::Container,
    ) -> <Self::ZoneExemplars as NamesContainer<TimezoneNamesCitiesOverrideV1, ()>>::Container;
    fn map_zone_exemplars_root(
        other: <M::ZoneExemplarsRoot as NamesContainer<TimezoneNamesCitiesRootV1, ()>>::Container,
    ) -> <Self::ZoneExemplarsRoot as NamesContainer<TimezoneNamesCitiesRootV1, ()>>::Container;
    fn map_zone_generic_long(
        other: <M::ZoneGenericLong as NamesContainer<TimezoneNamesGenericLongV1, ()>>::Container,
    ) -> <Self::ZoneGenericLong as NamesContainer<TimezoneNamesGenericLongV1, ()>>::Container;
    fn map_zone_generic_short(
        other: <M::ZoneGenericShort as NamesContainer<TimezoneNamesGenericShortV1, ()>>::Container,
    ) -> <Self::ZoneGenericShort as NamesContainer<TimezoneNamesGenericShortV1, ()>>::Container;
    fn map_zone_standard_long(
        other: <M::ZoneStandardLong as NamesContainer<TimezoneNamesStandardLongV1, ()>>::Container,
    ) -> <Self::ZoneStandardLong as NamesContainer<TimezoneNamesStandardLongV1, ()>>::Container;
    fn map_zone_specific_long(
        other: <M::ZoneSpecificLong as NamesContainer<TimezoneNamesSpecificLongV1, ()>>::Container,
    ) -> <Self::ZoneSpecificLong as NamesContainer<TimezoneNamesSpecificLongV1, ()>>::Container;
    fn map_zone_specific_short(
        other: <M::ZoneSpecificShort as NamesContainer<TimezoneNamesSpecificShortV1, ()>>::Container,
    ) -> <Self::ZoneSpecificShort as NamesContainer<TimezoneNamesSpecificShortV1, ()>>::Container;
    fn map_metazone_lookup(
        other: <M::MetazoneLookup as NamesContainer<TimezoneMetazonePeriodsV1, ()>>::Container,
    ) -> <Self::MetazoneLookup as NamesContainer<TimezoneMetazonePeriodsV1, ()>>::Container;

}

A trait for a DateTimeNamesMarker that can be created from a more specific one, M.

This trait is blanket-implemented on all field sets that are more general than M.

Examples

Example pairs of field sets where the trait is implemented:

use icu::datetime::fieldsets::enums::CompositeDateTimeFieldSet;
use icu::datetime::fieldsets::enums::CompositeFieldSet;
use icu::datetime::fieldsets::enums::DateFieldSet;
use icu::datetime::fieldsets::enums::TimeFieldSet;
use icu::datetime::fieldsets::T;
use icu::datetime::fieldsets::YMD;
use icu::datetime::scaffold::DateTimeNamesFrom;
use icu::datetime::scaffold::DateTimeNamesMarker;

fn is_trait_implemented<Source, Target>()
where
    Source: DateTimeNamesMarker,
    Target: DateTimeNamesFrom<Source>,
{
}

is_trait_implemented::<YMD, DateFieldSet>();
is_trait_implemented::<YMD, CompositeDateTimeFieldSet>();
is_trait_implemented::<YMD, CompositeFieldSet>();
is_trait_implemented::<T, TimeFieldSet>();
is_trait_implemented::<T, CompositeDateTimeFieldSet>();
is_trait_implemented::<T, CompositeFieldSet>();
is_trait_implemented::<DateFieldSet, CompositeDateTimeFieldSet>();
is_trait_implemented::<DateFieldSet, CompositeFieldSet>();
is_trait_implemented::<TimeFieldSet, CompositeDateTimeFieldSet>();
is_trait_implemented::<TimeFieldSet, CompositeFieldSet>();

Required Methods

fn map_year_names( other: <M::YearNames as NamesContainer<YearNamesV1, YearNameLength>>::Container, ) -> <Self::YearNames as NamesContainer<YearNamesV1, YearNameLength>>::Container

fn map_month_names( other: <M::MonthNames as NamesContainer<MonthNamesV1, MonthNameLength>>::Container, ) -> <Self::MonthNames as NamesContainer<MonthNamesV1, MonthNameLength>>::Container

fn map_weekday_names( other: <M::WeekdayNames as NamesContainer<WeekdayNamesV1, WeekdayNameLength>>::Container, ) -> <Self::WeekdayNames as NamesContainer<WeekdayNamesV1, WeekdayNameLength>>::Container

fn map_day_period_names( other: <M::DayPeriodNames as NamesContainer<DayPeriodNamesV1, DayPeriodNameLength>>::Container, ) -> <Self::DayPeriodNames as NamesContainer<DayPeriodNamesV1, DayPeriodNameLength>>::Container

fn map_zone_essentials( other: <M::ZoneEssentials as NamesContainer<TimezoneNamesEssentialsV1, ()>>::Container, ) -> <Self::ZoneEssentials as NamesContainer<TimezoneNamesEssentialsV1, ()>>::Container

fn map_zone_locations( other: <M::ZoneLocations as NamesContainer<TimezoneNamesLocationsOverrideV1, ()>>::Container, ) -> <Self::ZoneLocations as NamesContainer<TimezoneNamesLocationsOverrideV1, ()>>::Container

fn map_zone_locations_root( other: <M::ZoneLocationsRoot as NamesContainer<TimezoneNamesLocationsRootV1, ()>>::Container, ) -> <Self::ZoneLocationsRoot as NamesContainer<TimezoneNamesLocationsRootV1, ()>>::Container

fn map_zone_exemplars( other: <M::ZoneExemplars as NamesContainer<TimezoneNamesCitiesOverrideV1, ()>>::Container, ) -> <Self::ZoneExemplars as NamesContainer<TimezoneNamesCitiesOverrideV1, ()>>::Container

fn map_zone_exemplars_root( other: <M::ZoneExemplarsRoot as NamesContainer<TimezoneNamesCitiesRootV1, ()>>::Container, ) -> <Self::ZoneExemplarsRoot as NamesContainer<TimezoneNamesCitiesRootV1, ()>>::Container

fn map_zone_generic_long( other: <M::ZoneGenericLong as NamesContainer<TimezoneNamesGenericLongV1, ()>>::Container, ) -> <Self::ZoneGenericLong as NamesContainer<TimezoneNamesGenericLongV1, ()>>::Container

fn map_zone_generic_short( other: <M::ZoneGenericShort as NamesContainer<TimezoneNamesGenericShortV1, ()>>::Container, ) -> <Self::ZoneGenericShort as NamesContainer<TimezoneNamesGenericShortV1, ()>>::Container

fn map_zone_standard_long( other: <M::ZoneStandardLong as NamesContainer<TimezoneNamesStandardLongV1, ()>>::Container, ) -> <Self::ZoneStandardLong as NamesContainer<TimezoneNamesStandardLongV1, ()>>::Container

fn map_zone_specific_long( other: <M::ZoneSpecificLong as NamesContainer<TimezoneNamesSpecificLongV1, ()>>::Container, ) -> <Self::ZoneSpecificLong as NamesContainer<TimezoneNamesSpecificLongV1, ()>>::Container

fn map_zone_specific_short( other: <M::ZoneSpecificShort as NamesContainer<TimezoneNamesSpecificShortV1, ()>>::Container, ) -> <Self::ZoneSpecificShort as NamesContainer<TimezoneNamesSpecificShortV1, ()>>::Container

fn map_metazone_lookup( other: <M::MetazoneLookup as NamesContainer<TimezoneMetazonePeriodsV1, ()>>::Container, ) -> <Self::MetazoneLookup as NamesContainer<TimezoneMetazonePeriodsV1, ()>>::Container

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.