icu_datetime::scaffold

Trait NamesContainer

pub trait NamesContainer<M: DynamicDataMarker, Variables>: UnstableSealed
where
    Variables: PartialEq + Copy + Debug,
{
    type Container: MaybePayload<M, Variables> + Debug + Clone;
}

Trait that associates a container for a payload parameterized by the given variables.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Associated Types

type Container: MaybePayload<M, Variables> + Debug + Clone