icu_datetime::scaffold

Trait TimeMarkers

pub trait TimeMarkers: UnstableSealed {
    type HourInput: IntoOption<Hour>;
    type MinuteInput: IntoOption<Minute>;
    type SecondInput: IntoOption<Second>;
    type NanosecondInput: IntoOption<Nanosecond>;
    type TimeSkeletonPatternsV1: DataMarker<DataStruct = PackedPatterns<'static>>;
    type DayPeriodNamesV1: DataMarker<DataStruct = LinearNames<'static>>;
}

A trait associating types for time formatting (input types and data markers).

This is a sealed trait implemented on field set markers.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Associated Types

type HourInput: IntoOption<Hour>

Marker for resolving the day-of-month input field.

type MinuteInput: IntoOption<Minute>

Marker for resolving the day-of-week input field.

type SecondInput: IntoOption<Second>

Marker for resolving the day-of-year input field.

type NanosecondInput: IntoOption<Nanosecond>

Marker for resolving the any-calendar-kind input field.

type TimeSkeletonPatternsV1: DataMarker<DataStruct = PackedPatterns<'static>>

Marker for loading time skeleton patterns.

type DayPeriodNamesV1: DataMarker<DataStruct = LinearNames<'static>>

Marker for loading day period names.