Struct icu_datetime::DateTimeFormatter

pub struct DateTimeFormatter(_, _);

DateTimeFormatter is a formatter capable of formatting date/times from any calendar, selected at runtime. For the difference between this and TypedDateTimeFormatter, please read the crate root docs.

When constructed, it uses data from the data provider, selected locale and provided options to collect all data necessary to format any dates into that locale.

For that reason, one should think of the process of formatting a date in two steps - first, a computational heavy construction of DateTimeFormatter, and then fast formatting of DateTime data using the instance.

Examples

use icu::calendar::DateTime;
use icu::datetime::{options::length, DateTimeFormatter};
use icu::locid::Locale;
use std::str::FromStr;

let provider = icu_testdata::get_provider();

let mut options = length::Bag::from_date_time_style(length::Date::Medium, length::Time::Short);

let dtf = DateTimeFormatter::try_new_with_buffer_provider(&provider, &Locale::from_str("en-u-ca-gregory").unwrap().into(), &options.into())
    .expect("Failed to create TypedDateTimeFormatter instance.");

let datetime = DateTime::new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");
let any_datetime = datetime.to_any();

let value = dtf.format_to_string(&any_datetime).expect("calendars should match");
assert_eq!(value, "Sep 1, 2020, 12:34 PM");

Since this works with AnyCalendar, you can use DateTime with AnyCalendar to have a date in a runtime-selected calendar:

use icu::calendar::{AnyCalendar, AnyCalendarKind, DateTime, types::Time};
use icu::datetime::{options::length, DateTimeFormatter};
use icu::locid::Locale;

let provider = icu_testdata::get_provider();

let locale = Locale::from_str("en-u-ca-japanese").unwrap(); // English with the Japanese calendar

let calendar = AnyCalendar::try_new_with_buffer_provider(&provider, (&locale).try_into().unwrap())
                   .expect("constructing AnyCalendar failed");
let calendar = Rc::new(calendar); // Avoid cloning it

// manually construct a datetime in this calendar
let manual_time = Time::try_new(12, 33, 12, 0).expect("failed to construct Time");
// construct from era code, year, month code, day, time, and a calendar
// This is March 28, 15 Heisei
let manual_datetime = DateTime::new_from_codes("heisei".parse().unwrap(), 15, "M03".parse().unwrap(), 28,
                                               manual_time, calendar.clone())
                    .expect("Failed to construct DateTime manually");

// construct another datetime by converting from ISO
let iso_datetime = DateTime::new_iso_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct ISO DateTime.");
let iso_converted = iso_datetime.to_calendar(calendar);

let mut options = length::Bag::from_date_time_style(length::Date::Medium, length::Time::Short);

let dtf = DateTimeFormatter::try_new_with_buffer_provider(&provider, &locale.into(), &options.into())
    .expect("Failed to create TypedDateTimeFormatter instance.");

let manual_value = dtf.format_to_string(&manual_datetime).expect("calendars should match");
assert_eq!(manual_value, "Mar 28, 15 Heisei, 12:33 PM");

let converted_value = dtf.format_to_string(&iso_converted).expect("calendars should match");
assert_eq!(converted_value, "Sep 1, 2 Reiwa, 12:34 PM");

This model replicates that of ICU and ECMA402.

Implementations

impl DateTimeFormatter

pub fn try_new_with_any_provider<P>(
data_provider: &P,
locale: &DataLocale,
options: &DateTimeFormatterOptions
) -> Result<Self, DateTimeFormatterError> where
P: AnyProvider,

Construct a new DateTimeFormatter from a data provider that implements AnyProvider.

The provider must be able to provide data for the following keys: datetime/symbols@1, datetime/timelengths@1, datetime/timelengths@1, datetime/symbols@1, datetime/skeletons@1, datetime/week_data@1, and plurals/ordinals@1.

Furthermore, based on the type of calendar used, one of the following data keys may be necessary:

u-ca-japanese (Japanese calendar): calendar/japanese@1

pub fn try_new_with_buffer_provider<P>(
data_provider: &P,
locale: &DataLocale,
options: &DateTimeFormatterOptions
) -> Result<Self, DateTimeFormatterError> where
P: BufferProvider,

Construct a new DateTimeFormatter from a data provider that implements BufferProvider.

The provider must be able to provide data for the following keys: datetime/symbols@1, datetime/datelengths@1, datetime/timelengths@1, datetime/symbols@1, datetime/skeletons@1, datetime/week_data@1, and plurals/ordinals@1.

Furthermore, based on the type of calendar used, one of the following data keys may be necessary:

u-ca-japanese (Japanese calendar): calendar/japanese@1

Examples

use icu::calendar::DateTime;
use icu::datetime::{options::length, DateTimeFormatter};
use icu::locid::Locale;
use icu_provider::any::DynamicDataProviderAnyMarkerWrap;
use std::str::FromStr;

let provider = icu_testdata::get_provider();

let mut options = length::Bag::from_date_time_style(length::Date::Medium, length::Time::Short);
let locale = Locale::from_str("en-u-ca-gregory").unwrap();

let dtf = DateTimeFormatter::try_new_with_buffer_provider(&provider, &locale.into(), &options.into())
    .expect("Failed to create TypedDateTimeFormatter instance.");

let datetime = DateTime::new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");
let any_datetime = datetime.to_any();

let value = dtf.format_to_string(&any_datetime).expect("calendars should match");
assert_eq!(value, "Sep 1, 2020, 12:34 PM");

pub fn try_new_unstable<P>(
data_provider: &P,
locale: &DataLocale,
options: &DateTimeFormatterOptions
) -> Result<Self, DateTimeFormatterError> where
P: DataProvider<TimeSymbolsV1Marker> + DataProvider<TimeLengthsV1Marker> + DataProvider<DateSkeletonPatternsV1Marker> + DataProvider<OrdinalV1Marker> + DataProvider<WeekDataV1Marker> + DataProvider<DecimalSymbolsV1Marker> + DataProvider<GregorianDateLengthsV1Marker> + DataProvider<BuddhistDateLengthsV1Marker> + DataProvider<JapaneseDateLengthsV1Marker> + DataProvider<JapaneseExtendedDateLengthsV1Marker> + DataProvider<CopticDateLengthsV1Marker> + DataProvider<IndianDateLengthsV1Marker> + DataProvider<EthiopicDateLengthsV1Marker> + DataProvider<GregorianDateSymbolsV1Marker> + DataProvider<BuddhistDateSymbolsV1Marker> + DataProvider<JapaneseDateSymbolsV1Marker> + DataProvider<JapaneseExtendedDateSymbolsV1Marker> + DataProvider<CopticDateSymbolsV1Marker> + DataProvider<IndianDateSymbolsV1Marker> + DataProvider<EthiopicDateSymbolsV1Marker> + DataProvider<JapaneseErasV1Marker> + DataProvider<JapaneseExtendedErasV1Marker> + ?Sized,

Construct a new DateTimeFormatter from a data provider that can provide all of the requested data.

This method is unstable, more bounds may be added in the future as calendar support is added. It is preferable to use a provider that implements DataProvider<D> for all D, and ensure data is loaded as appropriate. The Self::try_new_with_buffer_provider(), Self::try_new_with_any_provider() constructors may also be used if compile stability is desired.

Examples

use icu::calendar::DateTime;
use icu::datetime::{options::length, DateTimeFormatter};
use icu::locid::Locale;
use icu_provider::any::DynamicDataProviderAnyMarkerWrap;
use std::str::FromStr;

let provider = icu_testdata::get_provider();

let mut options = length::Bag::from_date_time_style(length::Date::Medium, length::Time::Short);
let locale = Locale::from_str("en-u-ca-gregory").unwrap();

let dtf = DateTimeFormatter::try_new_unstable(&provider, &locale.into(), &options.into())
    .expect("Failed to create TypedDateTimeFormatter instance.");

let datetime = DateTime::new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");
let any_datetime = datetime.to_any();

let value = dtf.format_to_string(&any_datetime).expect("calendars should match");
assert_eq!(value, "Sep 1, 2020, 12:34 PM");

pub fn try_from_date_and_time(
date: DateFormatter,
time: TimeFormatter
) -> Result<Self, DateTimeFormatterError>

Constructor that takes a TimeFormatter and DateFormatter and combines them into a DateTimeFormatter. Prefer using one of the other constructors if possible.

Examples

use icu::calendar::DateTime;
use icu::datetime::{options::length, DateTimeFormatter, DateFormatter, TimeFormatter};
use icu::locid::{locale, Locale};
use icu_provider::any::DynamicDataProviderAnyMarkerWrap;
use std::str::FromStr;

let provider = icu_testdata::get_provider();

let length = length::Date::Medium;
let locale = Locale::from_str("en-u-ca-gregory").unwrap();

let df = DateFormatter::try_new_with_buffer_provider(&provider, &locale.into(), length)
    .expect("Failed to create TypedDateFormatter instance.");

let tf = TimeFormatter::try_new(
    &provider,
    &locale!("en").into(),
    length::Time::Short,
)
.expect("Failed to create TimeFormatter instance.");

let dtf = DateTimeFormatter::try_from_date_and_time(df, tf).unwrap();

let datetime = DateTime::new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");
let any_datetime = datetime.to_any();

let value = dtf.format_to_string(&any_datetime).expect("calendars should match");
assert_eq!(value, "Sep 1, 2020, 12:34 PM");

pub fn format<'l, T>(
&'l self,
value: &T
) -> Result<FormattedDateTime<'l>, DateTimeFormatterError> where
T: DateTimeInput<Calendar = AnyCalendar>,

Takes a DateTimeInput implementer and returns an instance of a FormattedDateTime that contains all information necessary to display a formatted date and operate on it.

This function will fail if the date passed in uses a different calendar than that of the AnyCalendar. Please convert dates before passing them in if necessary. This function will automatically convert and format dates that are associated with the ISO calendar.

pub fn format_to_write(
&self,
w: &mut impl Write,
value: &impl DateTimeInput<Calendar = AnyCalendar>
) -> Result<(), DateTimeFormatterError>

Takes a mutable reference to anything that implements Write trait and a DateTimeInput implementer and populates the buffer with a formatted value.

This function will fail if the date passed in uses a different calendar than that of the AnyCalendar. Please convert dates before passing them in if necessary. This function will automatically convert and format dates that are associated with the ISO calendar.

pub fn format_to_string(
&self,
value: &impl DateTimeInput<Calendar = AnyCalendar>
) -> Result<String, DateTimeFormatterError>

Takes a DateTimeInput implementer and returns it formatted as a string.

This function will fail if the date passed in uses a different calendar than that of the AnyCalendar. Please convert dates before passing them in if necessary. This function will automatically convert and format dates that are associated with the ISO calendar.

pub fn resolve_components(&self) -> Bag

Returns a components::Bag that represents the resolved components for the options that were provided to the DateTimeFormatter. The developer may request a certain set of options for a DateTimeFormatter but the locale and resolution algorithm may change certain details of what actually gets resolved.

Examples

use icu::calendar::Gregorian;
use icu::datetime::{
    options::{components, length},
    DateTimeFormatter, DateTimeFormatterOptions,
};
use icu::locid::Locale;
use std::str::FromStr;

let options = length::Bag::from_date_style(length::Date::Medium).into();

let provider = icu_testdata::get_provider();
let dtf = DateTimeFormatter::try_new_with_buffer_provider(
    &provider,
    &Locale::from_str("en-u-ca-gregory").unwrap().into(),
    &options
)
.expect("Failed to create TypedDateTimeFormatter instance.");

let mut expected_components_bag = components::Bag::default();
expected_components_bag.year = Some(components::Year::Numeric);
expected_components_bag.month = Some(components::Month::Short);
expected_components_bag.day = Some(components::Day::NumericDayOfMonth);

assert_eq!(dtf.resolve_components(), expected_components_bag);