Struct icu_datetime::TypedDateTimeFormatter

pub struct TypedDateTimeFormatter<C>(/* private fields */);

TypedDateTimeFormatter is a formatter capable of formatting date/times from a calendar selected at compile time. For the difference between this and DateTimeFormatter, please read the crate root docs.

When constructed, it uses data from the data provider, selected locale and provided options to collect all data necessary to format any dates into that locale.

For that reason, one should think of the process of formatting a date in two steps - first, a computational heavy construction of TypedDateTimeFormatter, and then fast formatting of DateInput data using the instance.

Examples

use icu::calendar::{DateTime, Gregorian};
use icu::datetime::{options::length, TypedDateTimeFormatter};
use icu::locid::locale;
use writeable::assert_writeable_eq;

let mut options = length::Bag::from_date_time_style(
    length::Date::Medium,
    length::Time::Short,
);

let dtf = TypedDateTimeFormatter::<Gregorian>::try_new(
    &locale!("en").into(),
    options.into(),
)
.expect("Failed to create TypedDateTimeFormatter instance.");

let datetime = DateTime::try_new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");

assert_writeable_eq!(dtf.format(&datetime), "Sep 1, 2020, 12:34โฏPM");

This model replicates that of ICU and ECMA402.

Implementations

impl<C: CldrCalendar> TypedDateTimeFormatter<C>

pub fn try_from_date_and_time( date: TypedDateFormatter<C>, time: TimeFormatter ) -> Result<Self, DateTimeError>

Constructor that takes a TimeFormatter and TypedDateFormatter and combines them into a TypedDateTimeFormatter.

Examples

use icu::calendar::Gregorian;
use icu::datetime::{
    options::length, TimeFormatter, TypedDateFormatter,
    TypedDateTimeFormatter,
};
use icu::locid::locale;

let tf = TimeFormatter::try_new_with_length(
    &locale!("en").into(),
    length::Time::Short,
)
.expect("Failed to create TimeFormatter instance.");
let df = TypedDateFormatter::<Gregorian>::try_new_with_length(
    &locale!("en").into(),
    length::Date::Short,
)
.expect("Failed to create TypedDateFormatter instance.");

TypedDateTimeFormatter::<Gregorian>::try_from_date_and_time(df, tf)
    .unwrap();

pub fn try_new( locale: &DataLocale, options: DateTimeFormatterOptions ) -> Result<Self, DateTimeError>where Baked: DataProvider<<C as CldrCalendar>::DateLengthsV1Marker> + DataProvider<<C as CldrCalendar>::DateSymbolsV1Marker>,

Constructor that takes a selected locale, then collects all compiled data necessary to format date and time values into the given locale.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Examples

use icu::calendar::{DateTime, Gregorian};
use icu::datetime::{options::length, TypedDateTimeFormatter};
use icu::locid::locale;
use writeable::assert_writeable_eq;

let options = length::Bag::from_date_time_style(
    length::Date::Medium,
    length::Time::Medium,
);

let dtf = TypedDateTimeFormatter::<Gregorian>::try_new(
    &locale!("en").into(),
    options.into(),
)
.unwrap();

let datetime =
    DateTime::try_new_gregorian_datetime(2022, 8, 31, 1, 2, 3).unwrap();

assert_writeable_eq!(dtf.format(&datetime), "Aug 31, 2022, 1:02:03โฏAM");

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized), locale: &DataLocale, options: DateTimeFormatterOptions ) -> Result<Self, DateTimeError>

A version of Self::try_new that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider( provider: &(impl BufferProvider + ?Sized), locale: &DataLocale, options: DateTimeFormatterOptions ) -> Result<Self, DateTimeError>

A version of Self::try_new that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_unstable<D>( provider: &D, locale: &DataLocale, options: DateTimeFormatterOptions ) -> Result<Self, DateTimeError>where D: DataProvider<<C as CldrCalendar>::DateSymbolsV1Marker> + DataProvider<<C as CldrCalendar>::DateLengthsV1Marker> + DataProvider<TimeSymbolsV1Marker> + DataProvider<TimeLengthsV1Marker> + DataProvider<DecimalSymbolsV1Marker> + DataProvider<OrdinalV1Marker> + DataProvider<WeekDataV1Marker> + ?Sized,

A version of Self::try_new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn format<'l, T>(&'l self, value: &T) -> FormattedDateTime<'l>where T: DateTimeInput<Calendar = C>,

Takes a DateTimeInput implementer and returns an instance of a FormattedDateTime that contains all information necessary to display a formatted date and operate on it.

Examples

use icu::calendar::{DateTime, Gregorian};
use icu::datetime::TypedDateTimeFormatter;
use writeable::assert_writeable_eq;
use icu::locid::locale;
let dtf = TypedDateTimeFormatter::<Gregorian>::try_new(&locale!("en").into(), options.into())
    .expect("Failed to create TypedDateTimeFormatter instance.");

let datetime = DateTime::try_new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");

assert_writeable_eq!(dtf.format(&datetime), "12:34:28โฏPM");

pub fn format_to_string( &self, value: &impl DateTimeInput<Calendar = C> ) -> String

Takes a DateTimeInput implementer and returns it formatted as a string.

Examples

use icu::calendar::{DateTime, Gregorian};
use icu::datetime::TypedDateTimeFormatter;
use icu::locid::locale;
let dtf = TypedDateTimeFormatter::<Gregorian>::try_new(&locale!("en").into(), options.into())
    .expect("Failed to create TypedDateTimeFormatter instance.");

let datetime = DateTime::try_new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");

assert_eq!(dtf.format_to_string(&datetime), "12:34:28โฏPM");