Module icu_datetime::time_zone

A formatter specifically for the time zone.

Structs

ExemplarCityFormat 🔒

GenericLocationFormat 🔒

GenericNonLocationLongFormat 🔒

GenericNonLocationShortFormat 🔒

Iso8601Format 🔒

LocalizedGmtFormat 🔒

SpecificNonLocationLongFormat 🔒

SpecificNonLocationShortFormat 🔒

TimeZoneDataPayloads 🔒

A container contains all data payloads for CustomTimeZone.

TimeZoneFormatter

TimeZoneFormatter is available for users who need to separately control the formatting of time zones. Note: most users might prefer ZonedDateTimeFormatter, which includes default time zone formatting according to the calendar.

TimeZoneFormatterOptions

A bag of options to define how time zone will be formatted.

Enums

FallbackFormat

An enum for time zone fallback formats.

FallbackTimeZoneFormatterUnit 🔒

IsoFormat

Determines which ISO-8601 format should be used to format a GmtOffset.

IsoMinutes

Whether the minutes field should be optional or required in ISO-8601 format.

IsoSeconds

Whether the seconds field should be optional or excluded in ISO-8601 format.

TimeZoneFormatterUnit 🔒

ZeroPadding 🔒

Whether a field should be zero-padded in ISO-8601 format.

Traits

FormatTimeZone 🔒

FormatTimeZoneWithFallback 🔒

Functions

format_offset_hours 🔒

format_offset_minutes 🔒

format_offset_seconds 🔒

format_time_segment 🔒

Formats a time segment with optional zero-padding.

load 🔒

Loads a resource into its destination if the destination has not already been filled.