Trait icu_datetime::CldrCalendar

pub trait CldrCalendar: InternalCldrCalendar {
    type DateSymbolsV1Marker: KeyedDataMarker<Yokeable = DateSymbolsV1<'static>>;
    type DateLengthsV1Marker: KeyedDataMarker<Yokeable = DateLengthsV1<'static>>;

    const DEFAULT_BCP_47_IDENTIFIER: Value;

    // Provided method
    fn is_identifier_allowed_for_calendar(value: &Value) -> bool { ... }
}

A calendar that can be found in CLDR

New implementors of this trait will likely also wish to modify get_era_code_map() in the CLDR transformer to support any new era maps.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland.

Required Associated Types

type DateSymbolsV1Marker: KeyedDataMarker<Yokeable = DateSymbolsV1<'static>>

The data marker for loading symbols for this calendar.

type DateLengthsV1Marker: KeyedDataMarker<Yokeable = DateLengthsV1<'static>>

The data marker for loading length-patterns for this calendar.

Required Associated Constants

const DEFAULT_BCP_47_IDENTIFIER: Value

The Unicode BCP 47 identifier for the calendarโs skeleton If multiple BCP 47 identifiers work, this should be the default one when no others are provided

If is_identifier_allowed_for_calendar() is set, this only is used for loading skeletons data

Provided Methods

fn is_identifier_allowed_for_calendar(value: &Value) -> bool

Checks if a given BCP 47 identifier is allowed to be used with this calendar

By default, just checks against DEFAULT_BCP_47_IDENTIFIER

Object Safety

This trait is not object safe.