icu_datetime
Module input

A collection of utilities for representing and working with dates as an input to formatting operations.

Traits

DateInput

Representation of a formattable calendar date. Supports dates in any calendar system that uses solar days indexed by an era, year, month, and day.

DateTimeInput

A combination of a formattable calendar date and ISO time.

IsoTimeInput

Representation of a time of day according to ISO-8601 conventions. Always indexed from midnight, regardless of calendar system.

LocalizedDateTimeInput

A formattable calendar date and ISO time that takes the locale into account.

TimeZoneInput

Representation of a formattable time zone.