icu_datetime::scaffold

Trait ConvertCalendar

pub trait ConvertCalendar {
    type Converted<'a>: Sized;

    // Required method
    fn to_calendar<'a>(&self, calendar: &'a AnyCalendar) -> Self::Converted<'a>;
}

A type that can be converted into a specific calendar system.

Required Associated Types

type Converted<'a>: Sized

The converted type. This can be the same as the receiver type.

Required Methods

fn to_calendar<'a>(&self, calendar: &'a AnyCalendar) -> Self::Converted<'a>

Converts self to the specified AnyCalendar.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.