icu_datetime::scaffold

Trait DateInputMarkers

pub trait DateInputMarkers: UnstableSealed {
    type YearInput: IntoOption<YearInfo>;
    type MonthInput: IntoOption<MonthInfo>;
    type DayOfMonthInput: IntoOption<DayOfMonth>;
    type DayOfYearInput: IntoOption<DayOfYear>;
    type RataDieInput: IntoOption<RataDie>;
    type DayOfWeekInput: IntoOption<Weekday>;
}

A trait associating types for date formatting in any calendar (input types only).

This is a sealed trait implemented on field set markers.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Associated Types

type YearInput: IntoOption<YearInfo>

Marker for resolving the year input field.

type MonthInput: IntoOption<MonthInfo>

Marker for resolving the month input field.

type DayOfMonthInput: IntoOption<DayOfMonth>

Marker for resolving the day-of-month input field.

type DayOfYearInput: IntoOption<DayOfYear>

Marker for resolving the day-of-year input field.

type RataDieInput: IntoOption<RataDie>

Marker for resolving the day-of-year input field.

type DayOfWeekInput: IntoOption<Weekday>

Marker for resolving the day-of-week input field.