Module icu_datetime::time_zone

Structs

TimeZoneDataPayloads

A container contains all data payloads for TimeZone.

TimeZoneFormat

TimeZoneFormat uses data from the data provider, the selected Locale, and the provided TimeZoneFormatConfig to collect all data necessary to format time zones into that locale.

TimeZoneFormatOptions

A bag of options to define how time zone will be formatted.

Enums

FallbackFormat

An enum for fallback formats.

IsoFormat

Determines which ISO-8601 format should be used to format a GmtOffset.

IsoMinutes

Whether the minutes field should be optional or required in ISO-8601 format.

IsoSeconds

Whether the seconds field should be optional or excluded in ISO-8601 format.

TimeZoneFormatConfig

A config enum for initializing TimeZoneFormat.

ZeroPadding

Whether a field should be zero-padded in ISO-8601 format.